Changed Demo to keep running in background tabs
Added in a tiny, nearly silent audio effect to tell Chromium-based browsers to keep the light scheduler running in the background even when tabs are hidden.  This change is a response to updates to background tab resource throttling introduced to Chrome in January 2021 to disincentivize repeated polling as a means of checking for updates or animation changes when push notifications or animation frameworks would suffice.  You can read more about it here: https://developer.chrome.com/blog/timer-throttling-in-chrome-88/